fix default cargo loading time adjustment fixes #67
fix issue

add parent dir

refactor

import main from Bot

format

fix main

fix

rename file
