+++ Bot/config.py
@@ -2,9 +2,16 @@ import configparser
 import os
 from typing import Any, Callable, List
 
+from loguru import logger
+
+# CONSTANTS
+
 # IMPORTANT: this MUST be tested after any code change
 # If an extra sleep is introduced, it might affect the total estimated time
-DEFAULT_CARGO_LOADING_TIME_ADJUSTMENT = 420
+_DEFAULT_CARGO_LOADING_TIME_ADJUSTMENT = 350
+
+# Default value for expected warping time to belt, the time to sleep after warping
+_DEFAULT_WARPING_TIME = 70.0
 
 
 class ConfigHandler:
@@ -20,22 +27,17 @@ class ConfigHandler:
         return [key.strip() for key in keys.split(",") if key.strip()]
 
     def get_unlock_all_targets_key(self) -> str:
-        return str(self._get_setting("unlock_all_targets_key", self.config.get, ""))
+        return self._get_setting("unlock_all_targets_key", self.config.get, "")  # type: ignore
 
     def get_take_screenshots(self) -> bool:
         return self._get_boolean_setting("take_screenshots", False)
 
     def get_cargo_loading_time_adjustment(self) -> int:
-        return int(
-            self._get_setting(
-                "cargo_loading_time_adjustment",
-                self.config.getint,
-                DEFAULT_CARGO_LOADING_TIME_ADJUSTMENT,
-            )
-        )
+        default = _DEFAULT_CARGO_LOADING_TIME_ADJUSTMENT + self.get_warping_time()
+        return self._get_setting("cargo_loading_time_adjustment", self.config.getint, default)  # type: ignore
 
     def get_mining_runs(self) -> int:
-        return int(self._get_setting("mining_runs", self.config.getint, 0))
+        return self._get_setting("mining_runs", self.config.getint, 0)  # type: ignore
 
     def get_undock_coo(self) -> List[int]:
         return self._get_coo("undock_coo")
@@ -59,30 +61,22 @@ class ConfigHandler:
         return self._get_coo("mouse_reset_coo")
 
     def get_mining_hold(self) -> int:
-        return int(self._get_setting("mining_hold", self.config.getint, 0))
+        return self._get_setting("mining_hold", self.config.getint, 0)  # type: ignore
 
     def get_mining_yield(self) -> float:
-        mining_yield = self._get_setting("mining_yield", self.config.get, "0")
-        return float(mining_yield.replace(",", "."))
+        return self._get_setting("mining_yield", self.config.getfloat, 0.0)  # type: ignore
 
     def get_mining_reset_timer(self) -> int:
-        mining_reset_timer = self._get_setting(
-            "mining_reset_timer", self.config.get, "120"
-        )
-        return 2 + int(mining_reset_timer)
+        return 2 + self._get_setting("mining_reset_timer", self.config.getint, 120)  # type: ignore
 
     def get_warping_time(self) -> int:
-        warping_time = self._get_setting("warping_time", self.config.get, "70")
-        return int(warping_time.replace(",", "."))
+        return self._get_setting("warping_time", self.config.getfloat, _DEFAULT_WARPING_TIME)  # type: ignore
 
     def set_mining_runs(self, value: str) -> None:
         self._set_setting("mining_runs", value)
 
     def set_undock_coo(self, value: str) -> None:
         self._set_position("undock_coo", value)
-
-    def set_hardener_key(self, value: str) -> None:
-        self._set_setting("hardener_key", value)
 
     def set_warp_to_coo(self, value: str) -> None:
         self._set_position("warp_to_coo", value)
@@ -100,13 +94,10 @@ class ConfigHandler:
         self._set_position("mouse_reset_coo", value)
 
     def set_mining_hold(self, value: str) -> None:
-        self._set_setting("mining_hold", str(value))
+        self._set_setting("mining_hold", value)
 
     def set_mining_yield(self, value: str) -> None:
-        self._set_setting("mining_yield", str(value))
-
-    def set_mining_reset_timer(self, value: str) -> None:
-        self._set_setting("mining_reset_timer", str(value))
+        self._set_setting("mining_yield", value)
 
     def set_mining_coo(self, value: str) -> None:
         self._set_position("mining_coo", value)
@@ -119,6 +110,11 @@ class ConfigHandler:
         try:
             return self.config.getboolean("SETTINGS", key)
         except configparser.NoOptionError:
+            logger.warning(
+                "Couldn't read {key} from settings, using default: {fallback}",
+                key=key,
+                fallback=fallback,
+            )
             return fallback
 
     def _get_setting(
@@ -131,6 +127,11 @@ class ConfigHandler:
             else:
                 return conversion_func("SETTINGS", key)
         except (configparser.NoSectionError, configparser.NoOptionError):
+            logger.warning(
+                "Couldn't read {key} from settings, using default: {fallback}",
+                key=key,
+                fallback=fallback,
+            )
             return fallback
 
     def _get_coo(self, key: str) -> List[int]:
@@ -150,7 +151,7 @@ class ConfigHandler:
             ]
 
     def _set_setting(self, key: str, value: str) -> None:
-        self.config.set("SETTINGS", key, str(value))
+        self.config.set("SETTINGS", key, value)
 
     def _set_position(self, key: str, value: str) -> None:
-        self.config.set("POSITIONS", key, str(value))
+        self.config.set("POSITIONS", key, value)
