--- dummy_test.py
@@ -1,6 +1,0 @@
-def inc(x: int) -> int:
-    return x + 1
-
-
-def test_answer() -> None:
-    assert inc(4) == 5
