+++ mypy.ini
@@ -1,0 +1,2 @@
+[mypy]
+exclude = tests
