+++ tests/test_config_with_warping_time.properties
@@ -1,0 +1,19 @@
+[SETTINGS]
+mining_runs = 
+mining_hold = 
+mining_yield = 
+mining_reset_timer = 
+hardener_keys = 
+warping_time = 140
+unlock_all_targets_key = 
+take_screenshots = False
+
+[POSITIONS]
+undock_coo = 
+clear_cargo_coo = 
+target_one_coo = 
+target_two_coo = 
+mouse_reset_coo = 
+warp_to_coo = 
+mining_coo = 
+
